backend/requirements.txt
Flask
pymongo
redis
python-dotenv
flask-cors
